Improved the service layer
